@@ -6,6 +6,6 @@ namespace RevXApi.Library.DataAccess
 	public interface IUserData
 	{
 		void CreateUser(UserModel user);
-		List<UserModel> GetUserById(string Id);
+		UserModel GetUserById(string Id);
 	}
 }
