@@ -6,9 +6,9 @@ namespace RevXApi.Library.Models
 	{
 		public int Id { get; set; }
 		public string UserId { get; set; }
-
 		public DateTime StartDate { get; set; }
 		public DateTime? EndDate { get; set; }
 		public double Rate { get; set; }
+		public int? ProviderId { get; set; }
 	}
 }
