@@ -11,6 +11,6 @@
             <LayoutView Layout="@typeof(MainLayout)">
                 <p>Sorry, there's nothing at this address.</p>
             </LayoutView>
-        </CascadingAuthenticationState>
+		</CascadingAuthenticationState>
     </NotFound>
 </Router>
