@@ -1,8 +1,13 @@
 ﻿@page "/management"
 @inject IProviderEndpoint _providerEndpoint
+@inject IStudentEndpoint _studentEndpoint
+@inject IHourlyRateEndpoint _rateEndpoint
 @inject IBillingStatusEndpoint _billingStatusEndpoint
 @inject ISessionEndpoint _sessionEndpoint
 @inject IJSRuntime jsRuntime
+@inject NavigationManager NavManager
+@inject IUserEndpoint _userEndpoint
+
 
 <div class="container mt-3 mb-3">
     @*<h2>Management</h2>
@@ -11,17 +16,24 @@
             <ExcelUpload />
         </div>*@
 
-    <SettingsContainer Header="General">
-        <Collapsible HeaderText="Manage Students">
-            <StudentEditTable></StudentEditTable>
-        </Collapsible>
-        <Collapsible Height="250" HeaderText="Manage Providers">
-            <ProviderEditTable></ProviderEditTable>
-        </Collapsible>
-        <Collapsible Height="200" HeaderText="Manage Hourly Rates">
-            <HourlyRateEditTable></HourlyRateEditTable>
-        </Collapsible>
-    </SettingsContainer>
+    <CascadingValue Value="Providers">
+        <CascadingValue Value="Students">
+            <CascadingValue Value="Rates">
+                <SettingsContainer Header="General">
+                    <Collapsible Height="@(Students is not null ? 8.3 + (Students.Count * 3.4) : 20)" HeaderText="Manage Students">
+                        <StudentEditTable></StudentEditTable>
+                    </Collapsible>
+                    <Collapsible Height="@(Providers is not null ? 8.3 + (Providers.Count * 3.4) : 20)" HeaderText="Manage Providers">
+                        <ProviderEditTable></ProviderEditTable>
+                    </Collapsible>
+                    <Collapsible Height="@(Rates is not null ? 8.3 + (Rates.Count * 3.4) : 20)" HeaderText="Manage Hourly Rates">
+                        <HourlyRateEditTable></HourlyRateEditTable>
+                    </Collapsible>
+                </SettingsContainer>
+            </CascadingValue> 
+        </CascadingValue>
+    </CascadingValue>
+
 
 </div>
 
@@ -29,4 +41,31 @@
 
 @code {
     // height is 150px the base plus 50px for each entry
+
+    [CascadingParameter]
+    private Task<AuthenticationState> authenticationStateTask { get; set; }
+    private string userId { get; set; }
+
+    public List<ProviderModel> Providers { get; set; }
+    public List<StudentModel> Students { get; set; }
+    public List<HourlyRate> Rates { get; set; }
+
+    protected override async Task OnParametersSetAsync()
+    {
+        if (!( await authenticationStateTask ).User.Identity.IsAuthenticated)
+        {
+            await jsRuntime.InvokeVoidAsync("log", "User id was null redirecting to login");
+            NavManager.NavigateTo($"/login?url={NavManager.Uri.Replace(NavManager.BaseUri, "")}");
+        }
+        else
+        {
+            UserModel user = await _userEndpoint.GetCurrentUser();
+            userId = user.Id;
+            await jsRuntime.InvokeVoidAsync("log", JsonConvert.SerializeObject(user));
+        }
+
+        Providers = await _providerEndpoint.GetAll(userId);
+        Students = await _studentEndpoint.GetAll(userId);
+        Rates = await _rateEndpoint.GetAll(userId);
+    }
 }
