@@ -1,9 +1,11 @@
 ﻿@page "/reports"
 @inject IReportEndpoint _reportEndpoint
+@inject NavigationManager NavManager
+@inject IJSRuntime jsRuntime
 
 <div class="mt-5"></div>
 <div class="mt-5"></div>
-<StackedBarChart RowCount="6" Data="MonthlyIncome" />
+<SchoolAfterSchoolReport  />
 <div>
     <div class="mt-3" style=" display: flex; width: max-content; margin-left: auto; margin-right: auto;">
         <p style="font-weight: bold; line-height: 1.1;">This report was generated by:</p>
@@ -28,25 +30,20 @@
     private Task<AuthenticationState> authenticationStateTask { get; set; }
     private string userId { get; set; }
 
-    public List<IncomeReportModel> MonthlyIncome { get; set; }
 
-    public List<double> MonthlyNumbers { get; set; } = new();
     public List<string> PieTooltipInfo { get; set; } = new();
-
-    public List<int> Values { get; set; }
-
+    public List<double> MonthlyNumbers { get; set; } = new();
     protected override async Task OnParametersSetAsync()
     {
-        userId = ( await authenticationStateTask ).User.FindFirst(ClaimTypes.NameIdentifier).Value;
-        MonthlyIncome = await _reportEndpoint.GetMonthlyIncome(DateTime.Parse("2020-09-01"), DateTime.Parse("2021-11-30"), userId, "month");
-        MonthlyNumbers = new();
-        foreach (var item in MonthlyIncome)
+        if (!( await authenticationStateTask ).User.Identity.IsAuthenticated)
         {
-            item.Title = DateTime.Parse(item.Date).ToString("MMMM");
-            var n = ( item.SchoolPrimary + item.AfterSchoolPrimary + item.SchoolSecondary + item.AfterSchoolSecondary );
-            var t = $"{DateTime.Parse(item.Date).ToString("Y")}\n{n} Hours.";
-            PieTooltipInfo.Add(t);
-            MonthlyNumbers.Add(n);
+            await jsRuntime.InvokeVoidAsync("log", "User id was null redirecting to login");
+            NavManager.NavigateTo($"/login?url={NavManager.Uri.Replace(NavManager.BaseUri, "")}");
+        }
+        else
+        {
+            userId = ( await authenticationStateTask ).User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            await jsRuntime.InvokeVoidAsync("log", $"User id was: {userId}");
         }
     }
 }
