@@ -26,7 +26,7 @@
                             </div>
                             <div class="form-group col-md-6">
                                 <label for="Date">Date</label>
-                                <InputDate type="date" min="2021-01-01" class="form-control" @bind-Value="@FormSession.Date" />
+                                <InputDate TValue="DateTime" min="2021-01-01" class="form-control" @bind-Value="@FormSession.Date" />
                             </div>
                         </div>
                         <div class="form-row">
