@@ -70,6 +70,7 @@
     private Task<AuthenticationState> authenticationStateTask { get; set; }
     private string userId { get; set; }
 
+    [CascadingParameter]
     public List<ProviderModel> providers { get; set; }
 
     private ProviderModel additionalProvider { get; set; } = new();
@@ -77,7 +78,6 @@
     protected override async Task OnParametersSetAsync()
     {
         userId = ( await authenticationStateTask ).User.FindFirst(ClaimTypes.NameIdentifier).Value;
-        await RefreshProviderList();
     }
 
     public void YouClickedMe(ProviderModel provider)
